Another trial of shorthand methods for ranges.
Background:
“For” loop with (init; target; increment) is kind of verbose.
Especially reverses… (7; >-1; ++) 
What are all those magic numbers. -1?
The built-in foreach doesn’t accept an array literal, since literals can
only be used in initializers. Wut.

So this works and shows the obvious numbers, but is detracts from readability:
for (int rx : new int[] {5, 4, 3, 2, 1, 0}) {}

Extracting to a constant helps but the constant also detracts:
for (int rx : FIVE_TO_ZERO) {}

The range method with explicit arguments and readable and functional.
The implementation is ugly but it is obvious, and buried.