Use a Map instead of a List for checking the level.
The List used an initial 0 value so that levels were in sync with
their level number, so the initial 0 was never used. Except it looks a
lot like Level 0, which is used for no-level Enchantments. Plus a chunk
piece could have 0 text pixels, adding more opportunities for bugs.

Map initialization is a little ugly, but it will be stowed in a class
eventually.

With the Map, the pixel counts are directly associated with the levels,
no extras or off-by-ones.

First draft of getting the Map value looked like this:
level = LEVEL_PIXELS.get(lastPixelCount);
if (level == null)…

get() returns null if the key cannot be found, but when autoboxing Integers,
that null is converted to an int, which can’t be null, so Null Pointer Exception.
Or something like that. I switched to containsKey(), since two method calls
are required to get the job done anyway. And containsKey() reads better.

And, move namePixels and level declarations closer to their use.