Save each enchantment image.
Eventually the image won’t be needed as a file, but it’s currently
useful for comparing visually via the HTML table feature.

In ChunkerTest, filter out filenames that are not 7 characters long,
since those represent output (the individual chunks), not input (the
full popup).

Throw all exceptions from everywhere.