Move options initialization to static{}.
The options need to be initialized once for the class, not 
for the instance. That was duplicating the options each call.

Remove display of options.size — and removeOption(), since
options must be kept in the array as
a future image may have a different number of pixels.

Increase height of dialog to 400. Looks better.