Ignore output files, give template real name.
ebi.html is the new base template file.
index.html is the filled in template, used for uploads etc.
ebi_rows.html is the raw output from the program

These files were originally in an already ignored directory,
but for convenience of uploading and management, they’re better at root.