Replace pixelMap with spellVault.
The number of pixels is now stored in Spells, a special case of
Enchantments.

Use a multi-step initialization process to make it easier to copy/paste
a new list of spells from a previous run’s output.

Remove the properties support since everything is now driven by 
Spells in the code.

(jOptionPane just moved, no changes.)